drive-service-4/requirements.txt
# ---- Servidor Web ----
fastapi
uvicorn[standard]
python-dotenv
google-cloud-storage
google-api-python-client
google-auth-oauthlib
pydantic